Implement language update and production meta data insertion
The language meta data is a little bit special as it needs to be added
in the meta data and also at the root element as the xml:lang
attribute. Copying the root node requires extra care (and code).

The production meta data is also a bit special in that it needs to be
added when missing.